+++ docs/src/components/Example.js
@@ -41,7 +41,7 @@ const Example = ({
       headingSize={headingSize}
     >
       <LiveProvider code={code} scope={scope} theme={theme}>
-        <Box display="flex" direction="column" marginLeft={-2} marginRight={-2}>
+        <Box display="flex" direction="column" marginStart={-2} marginEnd={-2}>
           <Box padding={2} height="100%">
             <Box position="relative" padding={4} borderStyle="sm" rounding={2}>
               <LivePreview />