+++ packages/gestalt/src/Box.test.js
@@ -63,7 +63,7 @@ test('Box has correct marginStart and marginEnd when marginStart and marginEnd a
 });
 
 test('Box allows auto margin', () => {
-  expect(create(<Box marginRight="auto" />)).toMatchSnapshot();
+  expect(create(<Box marginEnd="auto" />)).toMatchSnapshot();
 });
 
 test('Box has correct class when opacity is 0.5', () => {